Move PerformanceResult/Summary to results.h/cc
